Add files via upload
This is what I've got so far. Somehow Jupyter Notebook on VM doesn't see the module named 'pillow', although it has been installed with the 'pip install pillow' command. We have an exam in less than 9 hours, so I hope that it's possible to postpone the deadline by the end of the day, so that I'll be able to finish the task.